Fix a space leak in learned template set
There was a list of templates being updated via concatenation and a
simple `nubBy` that, combined with a `lookup` that never really forced
the end of the list, leaked the template memory.

This patch changes the template cache to a strict HashSet that does not
suffer from the problem. The performance should be otherwise identical.

Signed-off-by: Jan Dvořák <jan.dvorak@singularita.net>